Fetch accurate floating-point values from SQLite:
The Perl module is fetching from SQLite the numerical value of an
account/sub-account, but the numbers are rounded instead of being
accurate.

I found a bit of information on it here:

https://dba.stackexchange.com/questions/62491/why-does-sqlite-return-incorrect-sum

In any case, this use of `printf` allows fetching the correct data.
I verified this on "production" data of my own financial accounts. :)

A test was updated, but no particular test was added for this. The test
was updated because the comparison was a string (which now included two
decimal digits) instead of a numeric comparison.